# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Short-Term Rental Platform API
  description: |
    Vendor-neutral API for property managers, hosts, and guests.
    Built with privacy-by-design and ethical AI principles.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://api-staging.example.com/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local Development
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        description: Optional refresh token payload. If omitted, server may use refresh cookie.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update current user profile
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # Properties
  /properties:
    get:
      tags: [Properties]
      summary: Search properties
      operationId: searchProperties
      parameters:
        - name: location
          in: query
          schema:
            type: string
          description: Location search (city, address, or coordinates)
        - name: checkIn
          in: query
          schema:
            type: string
            format: date
        - name: checkOut
          in: query
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: propertyType
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/PropertyType'
        - name: amenities
          in: query
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertySearchResponse'
    post:
      tags: [Properties]
      summary: Create new property listing
      operationId: createProperty
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreate'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /properties/{propertyId}:
    get:
      tags: [Properties]
      summary: Get property details
      operationId: getProperty
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Properties]
      summary: Update property
      operationId: updateProperty
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdate'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
    delete:
      tags: [Properties]
      summary: Delete property
      operationId: deleteProperty
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '204':
          description: Property deleted
  /properties/{propertyId}/availability:
    get:
      tags: [Properties]
      summary: Get property availability
      operationId: getAvailability
      parameters:
        - $ref: '#/components/parameters/PropertyId'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Availability calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
    patch:
      tags: [Properties]
      summary: Update availability
      operationId: updateAvailability
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityUpdate'
      responses:
        '200':
          description: Availability updated
  /properties/{propertyId}/pricing:
    get:
      tags: [Properties]
      summary: Get property pricing rules
      operationId: getPricing
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Pricing rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingRules'
  /properties/{propertyId}/quote:
    post:
      tags: [Properties]
      summary: Get booking quote
      operationId: getQuote
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Price quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'

  # Bookings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PropertyId:
      name: propertyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Cursor for pagination
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string

    # Auth
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 10
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        accessToken:
          type: string
        token:
          type: string
          deprecated: true
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    # User
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        phone:
          type: string

    # Property
    PropertyType:
      type: string
      enum:
        - apartment
        - house
        - condo
        - townhouse
        - cabin
        - cottage
        - villa
        - chalet
        - loft
        - studio
        - guesthouse
        - bnb
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        roomType:
          type: string
          enum: [entire_place, private_room, shared_room]
        address:
          $ref: '#/components/schemas/Address'
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        bedrooms:
          type: integer
        beds:
          type: integer
        bathrooms:
          type: number
        maxGuests:
          type: integer
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        houseRules:
          $ref: '#/components/schemas/HouseRules'
        pricing:
          $ref: '#/components/schemas/PropertyPricing'
        rating:
          type: number
        reviewCount:
          type: integer
        status:
          type: string
          enum: [draft, listed, unlisted, suspended]
        instantBook:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PropertyCreate:
      type: object
      required: [title, propertyType, address, bedrooms, beds, bathrooms, maxGuests]
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 5000
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        roomType:
          type: string
          enum: [entire_place, private_room, shared_room]
        address:
          $ref: '#/components/schemas/Address'
        bedrooms:
          type: integer
          minimum: 0
        beds:
          type: integer
          minimum: 1
        bathrooms:
          type: number
          minimum: 0.5
        maxGuests:
          type: integer
          minimum: 1
        amenities:
          type: array
          items:
            type: string
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, listed, unlisted]
    PropertySearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        pagination:
          $ref: '#/components/schemas/Pagination'
        totalCount:
          type: integer
    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    Coordinates:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        caption:
          type: string
        order:
          type: integer
        isCover:
          type: boolean
    HouseRules:
      type: object
      properties:
        checkInTime:
          type: string
        checkOutTime:
          type: string
        smokingAllowed:
          type: boolean
        petsAllowed:
          type: boolean
        partiesAllowed:
          type: boolean
    PropertyPricing:
      type: object
      properties:
        basePrice:
          type: number
        currency:
          type: string
        cleaningFee:
          type: number
        weekendPrice:
          type: number

    # Availability
    AvailabilityResponse:
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
        dates:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityDate'
    AvailabilityDate:
      type: object
      properties:
        date:
          type: string
          format: date
        status:
          type: string
          enum: [available, blocked, booked]
        price:
          type: number
        minNights:
          type: integer
    AvailabilityUpdate:
      type: object
      properties:
        dates:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              status:
                type: string
                enum: [available, blocked]
              price:
                type: number
    PricingRules:
      type: object
      properties:
        basePrice:
          type: number
        currency:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PricingRule'
    PricingRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ruleType:
          type: string
          enum: [base, seasonal, weekend, length_of_stay, early_bird, last_minute]
        adjustmentType:
          type: string
          enum: [fixed, percentage]
        adjustmentValue:
          type: number
        isActive:
          type: boolean
    QuoteRequest:
      type: object
      required: [checkIn, checkOut, guests]
      properties:
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guests:
          type: integer
          minimum: 1
    QuoteResponse:
      type: object
      properties:
        available:
          type: boolean
        pricing:
          type: object
          properties:
            nightlyRate:
              type: number
            nights:
              type: integer
            subtotal:
              type: number
            cleaningFee:
              type: number
            serviceFee:
              type: number
            taxes:
              type: number
            total:
              type: number
            currency:
              type: string

    # Booking
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
        hasMore:
          type: boolean
        limit:
          type: integer

    # Economic
...